requirements.txt
# requirements.txt
langchain==0.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
llama-cpp-python==0.2.6
chromadb==0.4.15
sentence-transformers==2.2.2
pydantic==2.5.2
huggingface_hub==0.19.4
transformers==4.36.0